+++ internal/handlers/auth.go
@@ -6,18 +6,14 @@ import (
 	"net/http"
 	"time"
 
+	"github.com/Cahllagerfeld/go-htmx-first-steps/internal/auth"
 	"github.com/Cahllagerfeld/go-htmx-first-steps/internal/domain"
-	"github.com/Cahllagerfeld/go-htmx-first-steps/pkg/auth"
 	"github.com/gorilla/sessions"
 	"github.com/labstack/echo-contrib/session"
 	"github.com/labstack/echo/v4"
 	"github.com/markbates/goth"
 	"github.com/markbates/goth/gothic"
 )
-
-type AuthHandler struct {
-	e *echo.Echo
-}
 
 var users = []*domain.User{}
 
@@ -26,11 +22,7 @@ const (
 	tokenDuration time.Duration = time.Hour * 24 * 7
 )
 
-func NewAuthHandler(e *echo.Echo) *AuthHandler {
-	return &AuthHandler{e: e}
-}
-
-func (*AuthHandler) AuthCallback(ctx echo.Context) error {
+func authCallbackHandler(ctx echo.Context) error {
 	provider := ctx.Param("provider")
 	user, err := gothic.CompleteUserAuth(ctx.Response(), ctx.Request().WithContext(context.WithValue(ctx.Request().Context(), providerKey, provider)))
 	if err != nil {
@@ -52,7 +44,7 @@ func (*AuthHandler) AuthCallback(ctx echo.Context) error {
 	return ctx.Redirect(http.StatusFound, "/")
 }
 
-func (*AuthHandler) Logout(ctx echo.Context) error {
+func logoutHandler(ctx echo.Context) error {
 	sess, _ := session.Get(auth.SessionName, ctx)
 	sess.Options = &sessions.Options{
 		MaxAge:   -1,
@@ -67,7 +59,7 @@ func (*AuthHandler) Logout(ctx echo.Context) error {
 	return ctx.Redirect(http.StatusFound, "/")
 }
 
-func (*AuthHandler) Login(ctx echo.Context) error {
+func loginHandler(ctx echo.Context) error {
 	provider := ctx.Param("provider")
 	if _, err := gothic.CompleteUserAuth(ctx.Response(), ctx.Request().WithContext(context.WithValue(ctx.Request().Context(), providerKey, provider))); err == nil {
 		return ctx.Redirect(http.StatusFound, "/")
