+++ internal/router/router.go
@@ -2,11 +2,15 @@ package router
 
 import (
 	"github.com/labstack/echo/v4"
+	"github.com/labstack/echo/v4/middleware"
 	"github.com/labstack/gommon/log"
 )
 
 func New() *echo.Echo {
 	e := echo.New()
+	e.Use(middleware.Recover())
+	e.Use(middleware.Logger())
+
 	e.Logger.SetLevel(log.DEBUG)
 	e.Static("/assets", "dist")
