--- pkg/handlers/home.go
@@ -1,23 +1,0 @@
-package handlers
-
-import (
-	"context"
-
-	"github.com/Cahllagerfeld/go-htmx-first-steps/pkg/view"
-	"github.com/labstack/echo/v4"
-)
-
-type HomeHandler struct {
-	e *echo.Echo
-}
-
-func NewHomeHandler(e *echo.Echo) *HomeHandler {
-	return &HomeHandler{
-		e: e,
-	}
-}
-
-func (h *HomeHandler) Index(c echo.Context) error {
-	component := view.Hello("Baui")
-	return component.Render(context.Background(), c.Response().Writer)
-}
