+++ view/about.templ
@@ -1,6 +1,6 @@
 package view
 
-import "github.com/Cahllagerfeld/go-htmx-first-steps/pkg/layout"
+import "github.com/Cahllagerfeld/go-htmx-first-steps/view/layout"
 
 templ About() {
 	@layout.Base() {
