+++ view/hello_templ.go
@@ -10,7 +10,7 @@ import "context"
 import "io"
 import "bytes"
 
-import "github.com/Cahllagerfeld/go-htmx-first-steps/pkg/layout"
+import "github.com/Cahllagerfeld/go-htmx-first-steps/view/layout"
 
 func Hello(name string) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, templ_7745c5c3_W io.Writer) (templ_7745c5c3_Err error) {
@@ -38,7 +38,7 @@ func Hello(name string) templ.Component {
 			var templ_7745c5c3_Var3 string
 			templ_7745c5c3_Var3, templ_7745c5c3_Err = templ.JoinStringErrs(name)
 			if templ_7745c5c3_Err != nil {
-				return templ.Error{Err: templ_7745c5c3_Err, FileName: `pkg/view/hello.templ`, Line: 6, Col: 20}
+				return templ.Error{Err: templ_7745c5c3_Err, FileName: `view/hello.templ`, Line: 6, Col: 20}
 			}
 			_, templ_7745c5c3_Err = templ_7745c5c3_Buffer.WriteString(templ.EscapeString(templ_7745c5c3_Var3))
 			if templ_7745c5c3_Err != nil {
