+++ view/layout/base.templ
@@ -4,7 +4,7 @@ templ Base() {
 	<html>
 		<head>
 			<script src="https://unpkg.com/htmx.org/dist/htmx.min.js"></script>
-			<link rel="stylesheet" href="/assets/tailwind.css" />
+			<link rel="stylesheet" href="/assets/tailwind.css"/>
 		</head>
 		<body hx-boost="true" class="flex flex-col mx-auto max-w-[1440px]">
 			{ children... }
